Refactor cheers - put each command in a folder
