+++ commands/other/cheers/cheers.command.js
@@ -1,0 +1,9 @@
+const {buildCommand} = LyraCore
+
+module.exports = buildCommand({
+    run: require("./cheers"),
+    name: 'cheers',
+    group: 'fun',
+    memberName: 'cheers',
+    description: 'Feeling Cheery?'
+});
