+++ commands/other/cheers/cheers.js
@@ -1,0 +1,5 @@
+const cheersMessage = require("./messages/cheers.message")
+
+module.exports = async function (message) {
+    await message.say(cheersMessage());
+}
