+++ commands/other/cheers/messages/cheers.message.js
@@ -1,0 +1,7 @@
+const {baseEmbed} = LyraCore
+
+module.exports = function () {
+    return baseEmbed()
+        .setTitle("Cheers! :tea:")
+        .setImage("https://i.gifer.com/m9V.gif")
+}
