+++ commands/other/rps/rps.js
