+++ commands/other/vote/vote.js
